Create server.c file; Create socket for server
We create the new 'server.c' file, and write the logic to create
a socket for the server. It is configured to use the IPv4 domain
and is of type TCP (since we are building a TCP server). For this,
we use the function `socket()` from `<sys/socket.h>`. If it fails,
the program terminates with an error message and failure exit code.