Bind server socket to address
The next step, after creating the socket, is to bind it to an
address. For this case, the address consists of a port on the
host machine.

To do this, we configure the server by creating a struct of type
`struct sockaddr_in` (defined in `<netinet/in.h>`). We set the
server to use IPv4 addresses, listen on port 8888 (hardcoded), and
bind it to localhost (`127.0.0.1`) using the `inet_addr()` function
(from `<arpa/inet.h>`). This ensures that we can only connect to the
server from the local machine. To allow us to connect from anywhere,
this line can be replaced with `htonl(INADDR_ANY)`, which binds the
socket to all available interfaces.

We bind the server using the `bind()` function, and pass the file
descriptor (FD) of the socket, as well as the address and size of
the struct. Here, we must typecast the struct's address from type
`struct sockaddr_in *` to `struct sockaddr *`:
```
bind(socket_fd, **(struct sockaddr *)&server**, sizeof server)
```

This is because `bind()` can bind other types of sockets as well
(e.g. Unix sockets), and takes a general `struct sockaddr *` as
a parameter. As we are using an IP socket (for TCP), we can use
the specific type `struct sockaddr_in` and later typecast it. The
`bind()` function can deduce and handle this appropriately.