Set server to listen for connections
We want the server to be able to receive connections from clients.
To enable this, we use the `listen()` function, which takes 2
arguments: the FD of the server socket, and the number of connection
requests that will be queued before further requests are refused.
We have set this to 5 for now.