Set server to accept connections made to it
Since the server can now listen for connection requests, we can
configure it to accept those requests as well. We define another
struct of type `struct sockaddr_in` to represent the client, and
use the `accept()` function to accept the connection. It takes as
arguments the server socket FD, as well as pointers to the struct
and its size. The struct address is to be typecast to `struct
sockaddr *` from `struct sockaddr_in *`, and the struct size address
from `int *` to `socklen_t *`. `accept()` returns an FD for the socket
assigned to the new client connection.

We can test this by running the server, and attempting a connection
using either `telnet localhost 8888` or `nc localhost 8888`. In both
cases, doing so will have the server print an additional message
containing the client's socket FD, before the program terminates.