Allow server to send data to client connections
Since we want to have full communication between the server and
the client, we can have the server send a message to any clients
upon connecting with it. We can define a message string and pass
it to the function `send()`, which takes as arguments the
destination socket, the message and its length, and any flags to
set (for our use case, we leave it as 0).

When a client connects to the server, he should see the message
'Hello connection!' displayed on the terminal before the server
terminates.

There are additional functions that may be used to send data, such
as `write()` (defined in `<unistd.h>`: equivalent to `send()` with
the `flags` parameter equal to 0), or `dprintf()` (defined in
`<stdio.h>`: called like `fprintf()`, but with the destination FD
as the first parameter instead of the filestream).