Allow server to receive data from client connections
As part of full communication between the server and clients, we
also want clients to be able to send messages to the server and
get a response back.

We define a message buffer to store messages from the client. We
use the function `recv()` to take a message from a source socket.
The function takes as arguments the socket FD, the message buffer
and the no. of bytes to read, and any flags set (again, we leave
this as 0).

We can test the return value to handle errors: 0
indicates the stream socket has shut down (i.e. client disconnected),
and -1 indicates an error. If neither occurs, we have the server
send a response quoting what the client said. We can thus verify
that the server is correctly responding to the client.