Allow server to maintain multiple connections
Previously, the server would listen on the port, and would accept
the first client that requested a connection. Once the client
sent a message and got a response back, the server would simply
abruptly terminate. This meant that other clients attempting to
connect would be stuck waiting, and then suddenly dropped. This
effectively restricted the server to a single client.

To fix this, we move the `accept()` call into a loop. For every
new client attempting to connect, the function would be run to try
and assign a new socket for it. This allows mutiple clients to
connect to the server, which can be tested with multiple terminal
instances running `$ telnet localhost 8888`.

NOTE: The server still can only service one client at a time, with
the rest in the wait queue. Only after the first client sends a
message to the server and gets a response can the server accept a
connection from the next client in the queue. This means that the
server functions in a single-threaded/serial manner.