Allow server to send and receive data multiple times
Before, the server would take one message from the client and
send one message back, and then do nothing. Since we want the
server and client to exchange multiple messages, we need to
change this behaviour.

We achieve this by moving the `recv()` block into an additional
loop. As long as the client sends messages to the server, the
server will reply to all of them.