Server protocol: READ command
We begin implementing the intended protocol for the server. According
to the protocol, a message is a single line of text terminated by a
newline. For this, we start with the READ message. The server should
recognize the following:
```
READ <key>
```

To parse each message after reading it, we pass it to `strtok()` to
split it at `" "`. If the first token is not equal to `"READ"`, the
server does nothing more and continues to the next message. Otherwise,
it calls `strtok()` again to get the second token.

For now, the mock database is an array of strings, and we pass the key
in the message as the index of the string to retrieve. To convert the
key token from a string to a number, we use `strtol()` (`atoi()` is
more prone to errors, such as returning `0` for failure even if `"0"`
was passed to it).