Improve server protocol for READ
we continue to work on the protocol for the server-client
communication. We define an array `commands[]` of the commands that
the server will accept. We remove the previous use of `strtok()` and
replace it with logic that uses `strcspn()` to find the first
newline (`'\n'`) character and replace it with a NUL (`'\0'`)
character.

We iterate over `commands[]` and compare it with the client's message.
If it matches `"READ"`, we invoke `recv()` again to get the key
expected by `READ`. To avoid repeating the error handling section,
we use `goto` to jump out of the entire `while recv()` loop.

We convert the key string entered into an integer using `atoi()`.
We then access `database[]` and return the corresponding value.

**NOTE**: The implementation is not yet complete for READ. There are
potential errors or edge cases that need to be handled:
- Arbitrary whitespace before/after READ breaks the logic.
- `client_message[]` keeps old characters on subsequent reads; could
be handled by using `read_size` or `strcspn()` to always turn its
`'\n'` to `NUL`
- `atoi()` is error-prone and not the best way to convert strings to
integers in C. For one, it returns 0 even if conversion cannot be
done.
- The string comparison is possibly inefficient by needlessly calling
`strcmp()` and doing so inside a for loop.