Improve READ command further
This commit significantly changes the existing implementation for
the READ command. It strives to make the input processing faster
and less inefficient as compared to the previous state.
1. It now checks if the client simply pressed the `<Return>`/`<Enter>`
key (which sends a single `'\n'` character from `netcat`.) By checking
the value of `read_size` (`1` in this case), it can quickly discard
this input and return to awaiting user input.
2. The `'\n'` or final character from the input is always changed to a
`NUL` (`'\0'`) character. Its position is found from the value of
`read_size`.
3. The input now quickly discards inputs that are definitely not from
among the limited set of protocol commands. By checking the first
character entered, it can quickly discard such inputs and return to
awaiting user input. (**NOTE**: This approach only works as long as
the commands all start with different letters. Additionally, this
fails for when arbitrary whitespace is entered before the command -
approach to this not yet finalized.)
4. The command entered is tentatively identified from the first
character, and a string comparison using `strcmp()` is done between
the command and the client input (same pitfalls from previous point).
If it matches, the program runs `recv()` again to get the key that
READ expects. (If it fails, it breaks out of the while loop to reach
the error-handling logic.)

NOTES:
- The key is converted to an `int` from `char *` using `atoi()`, since
the test database is an array of strings. The final database is
intended to be a key-value data store with string keys and values.
- If more than 255 characters (incl. the newline) are entered in a
single user input, only the first 255 characters are read (i.e.
`read_size = 255`). This does not cause a buffer overflow or
read-past error, because the logic always sets the final character
entered into `NUL`, which is determined from the value of `read_size`.
Thus, either the newline or the last character in the buffer
`client_message[]` is always set to `'\0'`. **However**, the extra
characters entered are buffered, and are immediately entered as
the input to the next invocation of `recv()`. This may cause the
logic to execute a second time even without further user input, and
may cause unexpected/unintuitive interaction.
	e.g. `AAA...<255 times>READ`:
		- The first run (`read_size = 255`) discards the input.
		- The second run (`read_size = 5`) successfully processes
		the input and gets the command `READ`, and the server now
		enters a state of expecting the next user input to be the
		key for READ, even though this is not reflected at the
		client's end.