Server protocol: WRITE command
We implement the next command of the server-client protocol: `WRITE`.
Here, the syntax is as follows:
```
WRITE
<key to write at>
:<value to write>
```

The logic for `WRITE` is similar to that of `READ`, but the `recv()`
call is made twice (once for the key, once for the value). We then
make an access to the required location and write the entered value
in the (mock) database.

We also restructure the logic that attempts to match the command
entered. We use a `switch` statement to map commands to their
first letters (to quickly process *most* incorrect inputs), and
use `strcmp()` only for those inputs with the same first character
as a command (for now, all commands have different first letters; if
this were to change in the future, each new command would have its logic
under the corresponding `case`.) For all *individual* commands, a
`break` statement is executed immediately after their logic is
executed (since only a single command may be entered at a time).

NOTE: For the mock database, there is no bounds-checking implemented
for the array. This should not be an issue when it is replaced with a
C++ `std::unordered_map<std::string, std::string>`, but it is something
to be aware of for now.