Rename `server.c` to `server.cpp`
We are porting over the existing built server application from C
to C++. This cannot be done in one single step, and will be spread
over a few commits to ensure no breakages. The first step in this
is changing the file extension.

NOTE: To ensure Git is made aware of a tracked file that has been
renamed, use:
```
$ git mv <oldname> <newname>
```

Otherwise, if the file is directly renamed, Git is tricked into
thinking that <oldfile> was deleted, and <newfile> was newly
created with x number of insertions. It isn't a major issue, but
it is better to use `git mv` for a cleaner history.