Update C functions and headers used
As part of migrating the server program from C to C++, we update
the functions and headers from C to C++ -style syntax. We first
update the C headers `<stdio.h>`, `<stdlib.h>`, `<string.h>` to
`<cstdio>`, `<cstdlib>`, `<cstring>`. This change ensures that
the standard functions from these headers are included under the
`std` namespace (it is guaranteed that they will be under `std`;
they may also be in the global namespace but this is not
guaranteed.)

In the same vein, we prefix all the standard functions used from
these headers with `std::`, to adhere to good C++ practices.

The functions and headers related to sockets and networking
(e.g. `<netinet/in.h>`, `bind()`) are left unchanged.

NOTE: The function `dprintf()`, despite coming from `<stdio.h>`,
is not prefixed. This is because it is a GNU extension.

Finally, we include `<unordered_map>` and `<string>` (the C++ header,
not to be confused with `<cstring>`). This is so we can define the
intended key-value store for the server:
`std::unordered_map<std::string, std::string> KV_DATASTORE`.