Server protocol: COUNT command
This is a straightforward command that returns the number of
key-value pairs currently in the datastore. An unordered map has
a method `.size()` that returns this value, and which we can
invoke to return the number to the client.

**NOTE**:
The logic for the `READ` command was *broken* *before* this commit.
The original line read:
```
value = KV_DATASTORE[key];
```
The issue is that using `operator[]` returns the value **if** the
key exists, and **creates** it otherwise (in our case, the value
was always `""`.) This meant that `READ` was NOT idempotent.

To fix this, we use the `.find()` method of a map. It returns an
iterator to the key-value pair if the key is present, and is equal
to `.end()` otherwise. Thus, we can invoke the method and test its
return value to get the value only if the key already exists in the
map.

The exact type of the iterator can be:

- Explicitly specified
	`std::unordered_map<std::string, std::string>::iterator it`
	This is cumbersome to write out, and will break if we ever
	change the type of `KV_DATASTORE`.

- Inferred
	`auto it`
	This is useful as the actual type is not particularly useful to
	us, only what it represents and what we can compare it to.
	However, it does not make it clear that it is an iterator,
	which should NOT be dereferenced if it is equal to `.end()`.

- Implicitly specified
	`decltype(KV_DATASTORE)::iterator it`
	This is a good option as it saves typing, allows it to work
	regardless of a change to the key/value types of `KV_DATASTORE`,
	and still spells out that it represents an iterator object. Note
	that `decltype` deduces type at compile-time, while `typeid` does
	so at runtime.