Server protocol: DELETE command
This implements the DELETE command, which removes a key from the
datastore. If the key exists, it should be removed, and if not,
it should do nothing. Any change should appropriately be reflected
when using other commands (e.g. `COUNT` after successful `DELETE`
should show a decrease of 1, and `READ` after successful `DELETE`
should have no value. There should be no change whatsoever after
an unsuccessful `DELETE`.)

We use the map method `.erase()` for this. It is an overloaded
method with different invocations. We invoke it to take as
argument `const Key&`; in this case, it takes an `std::string`.
This returns the number of elements erased, which we can test to
see if the key existed in the datastore or not, and send the client
the appropriate message.