Server protocol: END command
This implements the `END` command for the server-client message
protocol. When the client sends this message to the server, the
server closes the connection established with that client.

For this, we use the `close()` function from `<unistd.h>`. It takes
the file descriptor to close as argument, and closes it. This makes
that FD available for use elsewhere. An alternate method is
`shutdown()` from `<sys/socket.h>`, which has more specific
functionalities (such as closing only the write end, or only the
read end, or both).

The logic to terminate the `recv()` loop requires us to directly
set the value of `read_size` to `0` (if the client, at any time,
pressed `<Ctrl>+<C>`, this would itself cause `read_size` to take
the value `0`, but we need to handle the explicit use of `END`.)

As this is inside a `switch` inside a `while`, using a `goto` to
jump directly outside both was the most straightforward option, and
spared us having to rearrange the code flow logic. We jump to a
label `after_recv:` outside the `while (recv())` loop and `switch`,
which takes control to the block that handles `read_size <= 0`. As
we set `read_size = 0` before jumping, it gets handled there.

The entire `while (recv())` loop itself lies in an outer
`while (accept())` loop. This means that even if a client ends the
connection, the server still stays up and listens for any new
connections made after that client.

NOTE: The use of `goto` here looks to be safe and not confusing, as
the target is close by, and it does not jump over any variable
declarations or initializations.