Server protocol fixes
1. READ command
We make minor tweaks to the server response to a `READ` command
from the client. In the given specification, the server responds
with the value of the key if it exists, and `NULL` otherwise. We
make changes accordingly. We also remove the helper print messages
(`Requested key:` ...).

2. WRITE command
We make minor tweaks to the `WRITE` command so that the server
replies with `FIN` after a successful write to the key-value
datastore.

3. COUNT command
We tweak the logic so the server only responds with the number of
key-value pairs, and remove the helper print statement. We also
change the formatting flag from `%lu` to `%ld` to avoid any
potential quirks arising from unsigned under/overflow.

4. DELETE command
Upon receiving the `DELETE` command, the server is supposed to
respond with `NULL` if the key in question was not found, and
with `FIN` if it was found (and thus erased).