Incorporate changes from given boilerplate `main.cpp`
The given boilerplate file `main.cpp` contains a few pre-written
comments explaining the program usage, as well as logic to handle
command-line arguments. We bring those changes over to the file
`server.cpp` that we have worked on so far.

The executable now expects a single command-line argument that
specifies the port on which to run the server. This value is
passed as the argument to `htons()` when configuring the port
value of the server in `server.sin_port`.