Change server logic to take bulk input
In the original logic, the server would take commands sent by the
client one by one. The server would run `recv()` repeatedly to
get more inputs from the client, and would send a reply the
moment it matched a valid command input sequence. However, this
made it fail when taking a file piped into `netcat` as input, and
makes it impossible to write the entire input before sending it
to the server in one go.

To fix this, we make major changes to the logic used, while also
using more C++ -typical features. We now use a single `std::string`
variable `client_message` to store the entire client input. We
run a loop to check if it contains `END`, and if not, we `recv()`
data from the client into a `char buffer[255]`, and append the
newly entered data to `client_message`. To avoid having to search
the entire string `client_message` on every iteration, we store
a value `end_pos` that stores the position after the most
recently added substring of `client_message`.

From the final value of `client_message`, we initialize a
variable `std::istringstream input`. By doing this, we can use
the `>>` operator to efficiently (compared to directly from
`stdin`/`cin`). Plus, it automatically skips all whitespace in between
(space, tabs, newline, etc.) and we do not need to handle those
ourselves.
We also create a `std::ostringstream reply` that will hold the
final response we will eventually make to the client once all the
commands have been processed (again, it is more efficient rather
than multiple calls to `<<` to `cout`/`stdout`).

The rest of the logic is more or less the same, just tweaked to use
the stringstream objects. We read a token from `input` into `command`
and process it (and read another token for the key value), and then
write the result and a newline to `reply`.

Once the `"END"` command is processed, the server dispatches the final
reply to the client before terminating the connection.

NOTES:
- **IMPORTANT**: This program currently works as intended only when
performing step-by-step execution in GDB. When running it as-is, it
produces no output. **FIX ASAP**
- I haven't tested it yet, but the segment `client_message.find("END"`
*might* try and match a literal `END` inside some other input,
instead of the actual command.
- The `goto` used in the `if (read_size == 0)` segment jumps backwards,
which might be bad practice, as it jumps over variable initializations.