Fix bug relating to `END` command
Earlier, the server logic would run a search in a loop to see if the
client had sent the `END` command to signal no more input. However,
the implementation included a hack that attempted to reduce the
amount of searching to be done by starting the search from near
the end of the string `client_message` thus far.

The issue was that it would, for some inputs and not others, cause
the server to be frozen in the `recv()` loop even after an `END`
command had been sent. This has been fixed by commenting out that
hack and having the search done on the entire string. The program
now works on both piped file input as well as manual input.

NOTE: The idea of improving the string `.find()` is still worth
implementing, but not a pressing concern at this moment.