Add .gitattributes file to handle line endings
The `test.sh` file provided is intended to be run within a Unix
environment/shell. However, if it is treated as a Windows file,
then it will contain carriage return (`CR`) characters that will
prevent it from executing in the shell.

To remedy this, we commit into version control a `.gitattributes`
file. In this, we declare that all `*.sh` files (the extension for
a shell script on Unix systems) are to *always* have LF line
endings, and never CRLF.