Begin converting serial singlethreaded server to multithreaded server
The current server is singlethreaded; it can only handle one client's
messages at a time, while other prospective clients are stuck waiting
in the `accept()` queue. Only once the current client sends the `END`
command and gets his response can the next client be `accept()`ed by
the server.

The intention is to migrate the existing server logic from its
current singlethreaded model to a multithreaded one by using the
POSIX Threads (pthreads) API. For this, we include the additional
header `<pthread.h>` in the source file. Additionally, we need to
link the pthread library against our executable, for which the
`-pthread` compiler flag (**NOT** `-lpthread`) is necessary.