Separate out client handling logic into function
As part of moving the server from a singlethreaded design to a
multithreaded one, we separate out the part of the logic that
handles interacting with a client. We move it into its own
function `handle_connection()`. This will help in subsequent work
when employing threads, as each new client can be handled by a
newly-spawned thread.

In order to keep it working, we move the `KV_DATASTORE` variable
out of `main()` and make it a global variable. Otherwise, it would
not be accessible from within `handle_connection()` unless we pass
a reference to it as an argument.

Additionally, the `accept_connections:` label with `goto` was
unnecessary in the original code, as it could have been expressed
with `continue` instead. The use of `goto` in this case was not
necessary.