Use threads to handle individual clients
In the singlethreaded server, when the server `accept()`ed a client,
all other clients trying to connect would be waiting in the queue
until the first client disconnects, at which point only the next
client in queue can be `accept()`ed.

To have the server be able to handle multiple clients simultaneously,
we change the server logic to spawn a new thread when a new client
attempts to connect, and have that thread handle the interaction with
the client while the server returns to listening for new clients.

We accomplish this using the Pthreads API provided by `<pthread.h>`.
We declare a thread variable of type `pthread_t`, and use the
function `pthread_create()` to create a new thread at the given
thread variable. The function can take a function pointer to execute
as well as the arguments it takes, and we can pass the client handler
function `handle_connection()` along with the client socket to
`pthread_create()`.

The Pthreads API requires that `pthread_create()` takes a pointer
to a function that takes a single `void *` argument and returns
`void *` (i.e. `void *(*)(void *)`), and the argument is also passed
as `void *`. We modify the signature of `handle_connection()`
accordingly, as well as the call to it. We pass the (`int`) client
socket by casting its address to `void *`, and then casting the
argument in `handle_connection()` back to `int *` before dereferencing
it to get the socket value.

Additionally, we make sure to join the new client thread to the main
thread by `pthread_join()`. This ensures that the main thread waits
for the client thread to finish its execution before it itself exits.
Otherwise, it is possible for the main thread to exit and have the
program abruptly terminate while the client thread is still occupied.
Testing the program without the `pthread_join()` caused either long
waiting with nothing happening, or errors like 'Bad file descriptor',
'Connection refused', or 'Connection reset by peer'.