Use mutex to protect critical section of client threads
As part of the available operations, clients may `WRITE` to or
`DELETE` from the key-value datastore. However, since this modifies
the state of the datastore, it makes it potentially cause data races.
Another thread may `READ` at the same time as a `WRITE`, and they
might return inconsistent results.

To enforce consistency, we protect the critical section of the client
threads using a mutex. As they modify the datastore, the parts of the
logic inside the `WRITE` and `DELETE` commands are parts of the
critical section.

We first declare a global mutex variable `MUTEX_FOR_KV_DATASTORE` of
type `pthread_mutex_t` that will be responsible for this. Inside the
main thread, we make use of `pthread_mutex_init()` to initialize the
variable, and we will finally destroy it with `pthread_mutex_destroy()`
before the main thread itself exits. These two are mandatory for proper
behaviour.

Inside `handle_connection()`, we need to lock the mutex immediately
before the relevant part of the `WRITE` and `DELETE` operations, and
immediately unlock it after. To lock the mutex, we call
`pthread_mutex_lock()`, and unlock it with `pthread_mutex_unlock()`.

- In `WRITE`, we lock it before `KV_DATASTORE[key] = value`, and unlock
it after.

- In `DELETE`, we lock it before `KV_DATASTORE.erase(key)`, and unlock
it after.

It should be noted that using mutexes ensures the consistency of the
datastore, but may impact performance. In some cases, it can cause
that portion of the program flow to be effectively serialized.

Note: The array of command strings `commands[]` was removed. It was
unused for some time (it was written for an earlier iteration of the
server command matching logic) and was no longer necessary.