Enforce LF line ending on automatic test template and expected output
It seems that the program is intended to be built, run, and tested on
a Unix environment, since the test runner is a Bash script that
makes calls to Unix system utilities (`ls`, etc.). Thus, it makes
sense to enforce LF line endings for the files:
- `tests/inputs/ALL_SUB.txt`
- `tests/expected_outputs/ALL_SUB.txt`

as the former is used to generate the needed inputs for test cases,
while the latter is the expected output from all inputs. The stray
`'\r'` characters cause the output to also expect `'\r'` characters,
but the program does not output them as it is being run on a Unix
environment (WSL2). Additionally, they are not even retained in the
client input strings as the stream `<<` and `>>` operators ignore all
whitespace.

If need be, this change can be easily reverted later, and it is more
convenient than having to call `dos2unix` before every testing session.