Detach client threads upon spawning them
We modify the thread spawn logic to detach the threads as soon as
we create them and assign them to a client. This means that the
main thread no longer waits for that client thread to return before
it resumes its own execution. This was caused by us joining the client
thread (with `thread_join()`), and this made the 'multithreaded'
server effectively singlethreaded.

This rectifies the change in commit `24d9662`, whose message also
erroneously says that joining is needed here. Detaching the client
thread means that the main thread remains inside the `while` loop
and goes back to `accept()`ing any new connections.

We also re-arrange the `accept()` while loop:
- Previously, it would run forever as long as `accept() > 0`, i.e.
as long as the call succeeded. This meant that the server would stop
listening and terminate once it failed to accept a client.
- Now, it is run forever (`while(1)`) and simply `continue`s if it
fails to `accept()` a connection.

Note: The `pthread_exit(NULL)` is functionally equivalent to simply
returning from the thread.