Rewrite logic to instead directly create detached client threads
In Pthreads, a thread can be in one of 2 states - joinable or
detached. By default, a thread created with `pthread_create()` returns
a joinable thread. However, a joinable thread is assumed to be needed
elsewhere and that the caller will be waiting for it. Thus, it assumes
that `pthread_join()` will be executed on it.

The issue is that joinable threads occupy thread IDs and resources
that need to be returned to the caller before they can be reused. This
can result in unnecessary load, since we do not need to have the
threads be joined.

Thus, instead of creating a thread and then detaching it, we directly
create it to be a detached thread.

We declare a variable of type `pthread_attr_t`, initialize it, and
then set its detach state to be detached, instead of joinable. We can
then pass this attribute variable to the call to `pthread_create()` to
have any new threads created by detached.