Allocate memory instead of directly passing client socket argument
When a new client connection is established, the `accept()` function
returns a new socket file descriptor that the server uses to
communicate with the client. To do this, we originally passed the
FD value as an argument to `handle_connection()`.

However, in the multithreaded server, each thread should have its
own copy of the socket FD it is assigned to. This avoids race
conditions and other contradictory behaviour that may otherwise
come up. While testing the server, it already showed this, when the
server would simply freeze and do nothing on some occasions, and work
normally other times.

To fix this, we dynamically allocate memory to store the value of
the client socket FD, and pass a pointer to it to `handle_connection()`.
This allows each thread to separately retrieve the needed socket FD.
Once the client closes the connection, the thread frees the allocated
memory before terminating.

Additionally, we set the server socket to be reusable with the help of
`setsockopt()` before binding it to an address and port. This ensures
that the server can be bound to an address and port even when a
previous connection on that port is in the `TIME_WAIT` state (which is
a mechanism in TCP that keeps sockets open even after an application
shuts down the sockets. It helps ensure complete data transmission is
done between the server and the client, which is part of how TCP
works.) By making the socket reusable, it makes it much easier to
quickly test the server during development.