Switch server to using a 'thread pool'-like approach
The server used to handle clients dynamically - when a new client
made a connection, the server would spin up a new thread that would
handle the communication with the client, while the server could return
to listening for more clients. This approach worked well, but there
are some costs to it. Creating threads involves some overhead as each
thread maintains its own stack space, etc. This means that repeatedly
creating and destroying threads builds up this resource overhead,
especially when there is a large number of tasks involved.

A solution is to make use of a 'thread pool', which is a set of
pre-created threads. Instead of creating and destroying threads,
we can reuse these threads to handle multiple tasks.

We implement a basic thread pool as an array of threads that we
pre-create in a loop. Note that the threads are still created as
detached threads, as the main thread does not need to wait for the
client handler threads to finish.

We change the function assigned to the threads to be `thread_function()`.
This function simply locks the client queue with another mutex, and
unlocks it to get the first client socket, which it passes to the
`handle_connection()` function, before locking it again.

`handle_connection()` itself functions in exactly the same way; only
its function signature has changed.

We pre-define the number of threads kept in the pool using a macro
`NUM_THREADS`, which is set to a value of `5`, but which can be
overriden by a command-line flag `-DNUM_THREADS=<value>`.