Restructure server code into multiple translation units
The `Tcp_Server` class was previously written as a single large unit
in the main file itself. To split it into its interface and its
implementation, we create 2 new files: `Tcp_Server.cpp` and
`Tcp_Server.hpp`.
The header is where the class and its data members
are declared.
The source is where the methods are actually defined
against the header interface.

We can compile `Tcp_Server.cpp` into a standalone object file. We
can do the same for `main.cpp`, which now only needs to include
the server class interface in `Tcp_Server.hpp`. The final executable
can then be simply built by linking the object files together.

We modify the `Makefile`s accordingly. `./src/parallel/` has its
own, which produces the object files, and builds the final
executable in the `./bin/` folder. The root `Makefile` simply
calls the one for `./src/parallel/`.